export global
